Atualizando método divisão

+++ modulo1/src/aula2/Calculadora.java
@@ -10,7 +10,7 @@ public class Calculadora {
 		return n1 - n2;
 	}
 	
-	int dividir (int n1, int n2) {
+	float dividir (float n1, float n2) {
 		return n1 / n2;
 	}
